+++ packages/dumb/src/components/stats/time-in-range/time-in-range-chart/time-in-range-chart.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2025, Diabeloop
+ * Copyright (c) 2022-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,27 +25,26 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import Box from '@mui/material/Box'
+import { type BgType, type CbgRangeStatistics } from 'medical-domain'
 import React, { type FunctionComponent } from 'react'
+import { type BgPrefs } from '../../../../models/blood-glucose.model'
+import { CBGStatType } from '../../../../models/stats.model'
+import { StatLegendMemoized as StatLegend } from '../../stat-legend/stat-legend'
 import { CBGPercentageBarMemoized as CBGPercentageBar } from '../cbg-percentage-bar/cbg-percentage-bar'
 import { TimeInRangeTitleMemoized as TimeInRangeTitle } from '../time-in-range-title'
 import { useTimeInRangeChartHook } from './time-in-range-chart.hook'
-import { CBGStatType } from '../../../../models/stats.model'
-import { StatLegendMemoized as StatLegend } from '../../stat-legend/stat-legend'
-import Box from '@mui/material/Box'
-import { type BgPrefs } from '../../../../models/blood-glucose.model'
-import { type BgType, type CbgRangeStatistics } from 'medical-domain'
 
 interface TimeInRangeChartProps {
   bgType: BgType
   bgPrefs: BgPrefs
-  cbgStatType: CBGStatType
   data: CbgRangeStatistics
   days: number
   trackHoverFunc: (name: string) => void
 }
 
 const TimeInRangeChart: FunctionComponent<TimeInRangeChartProps> = (props) => {
-  const { bgPrefs, bgType, cbgStatType, data, days, trackHoverFunc } = props
+  const { bgPrefs, bgType, data, days, trackHoverFunc } = props
   const {
     annotations,
     cbgStatsProps,
@@ -57,7 +56,6 @@ const TimeInRangeChart: FunctionComponent<TimeInRangeChartProps> = (props) => {
     bgType,
     data,
     days,
-    type: cbgStatType,
     bgPrefs
   })