+++ CHANGELOG.md
@@ -15,6 +15,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3945 Adapting the second level of navigation to mobile format
 - YLP-4042 Display G7 Pairing code instead of Firmware version on Device view
 - YLP-4046 Improve the display of the parameter changes tooltip on Daily view
+- YLP-4047 Use smart rounding for TIR percentage values
 
 ### Engineering
 - YLP-3864 SBOM generation in pipeline